+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Features
 - add support for defining GraphQL interfaces and implementing it by object types
 - add support for extending object type classes
+- add support for implementing GraphQL interfaces without decorators duplication
 
 ## v0.5.0
 ### Features